Actual add the storefront file. Refactor inventoryclient

+++ dao/InventoryItem.go
@@ -8,5 +8,5 @@ type InventoryItem struct {
 	Description       string
 	Category          string
 	Price             float64
-	QuantityAvailable int
+	QuantityAvailable uint64
 }
